Module re_space_view_tensor::space_view_class

Structs

ColorMapping

How we map values to colors.

PerTensorState

SliceSelection

How we slice a given tensor

TensorSpaceView

TextureSettings ğŸ”’

Scaling, filtering, aspect ratio, etc for the rendered texture.

ViewTensorState

Enums

TextureScaling ğŸ”’

Should we scale the rendered texture, and if so, how?

Constants

MIN_SLIDER_WIDTH ğŸ”’

Statics

ONCE ğŸ”’

__SEEN_MESSAGES ğŸ”’

Functions

colormap_preview_ui ğŸ”’

Show the given colormap as a horizontal bar.

dimension_name ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

paint_axis_names ğŸ”’

paint_colormap_gradient ğŸ”’

paint_tensor_slice ğŸ”’

selected_tensor_slice

selectors_ui ğŸ”’

tensor_slice_ui ğŸ”’

tf_to_string ğŸ”’

view_tensor ğŸ”’