Function re_space_view_tensor::dimension_mapping::load_tensor_slice_selection_and_make_valid

pub fn load_tensor_slice_selection_and_make_valid(
    slice_selection: &ViewProperty<'_>,
    shape: &[TensorDimension]
) -> Result<TensorSliceSelection, DeserializationError>

Loads slice selection from blueprint and makes modifications (without writing back) such that it is valid for the given tensor shape.

This is a best effort function and will insert fallbacks as needed. Note that fallbacks are defined on the spot here and don’t use the component fallback system. We don’t need the fallback system here since we’re also not using generic ui either.

General rules for scrubbing the input data:

out of bounds dimensions and indices are clamped to valid

missing width/height is filled in if there’s at least 2 dimensions.